refactor: removed unecessary files and moved cluster to internals
