+++ internal/cluster/client.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/rqlite/rqlite/v8/auth"
 	"github.com/rqlite/rqlite/v8/rtls"
-	"github.com/tarungka/wire/cluster/proto"
+	"github.com/tarungka/wire/internal/cluster/proto"
 	command "github.com/tarungka/wire/internal/command/proto"
 	"github.com/tarungka/wire/internal/tcp"
 	"github.com/tarungka/wire/internal/tcp/pool"
